requirements.txt

PyJWT
flask==1.1.1
flask-cors==3.0.8
flask-sqlalchemy
flask-migrate
Flask-Injector==0.12.0
flask-inputs==0.3.0
flask-marshmallow==0.10.1
flask-jwt-extended==3.23.0
flask-script
marshmallow-sqlalchemy==0.19.0
connexion[swagger-ui]
bcrypt
flask-bcrypt
numpy
psycopg2-binary
# psycopg2
gunicorn
pandas